+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2023-06-22
+
+### Fixed
+
+- Godot custom data or meta data did not always have the right type (issue #13). This is now fixed.
+- For some properties concerning layers non-existing layers were not created. This is now fixed.
+
 ## [1.3.0] - 2023-06-18
 
 ### Added